fixing attribute migration for big integer

+++ Modules/Attribute/Database/Migrations/2020_03_05_071620_create_attribute_groups_table.php
@@ -14,12 +14,12 @@ class CreateAttributeGroupsTable extends Migration
     public function up()
     {
         Schema::create('attribute_groups', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('name');
             $table->string('slug');
             $table->integer('position')->nullable();
             $table->boolean('is_user_defined')->default(1);
-            $table->integer('attribute_family_id')->unsigned();
+            $table->bigInteger('attribute_family_id')->unsigned();
             $table->foreign('attribute_family_id')->references('id')->on('attribute_families')->onDelete('cascade');
         });
     }
